third_party/WebKit/LayoutTests/http/tests/devtools/console/console-preserve-scroll-expected.txt

CONSOLE WARNING: line 3078: Element.createShadowRoot is deprecated and will be removed in M73, around March 2019. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 for more details.
Tests that console preserves scroll position when switching away.

Message count: 100
Console scrollTop: 10

Running: testSwitchToAnotherPanel
Panel sources was opened.
Panel console was opened.
Console scrollTop: 10

Running: testClickLinkToRevealAnotherPanel
Panel sources was opened.
Panel console was opened.
Console scrollTop: 10

Running: testConsolePanelToDrawer
Drawer panel set to console-view
Panel sources was opened.
Console scrollTop: 10
Panel console was opened.
Console scrollTop: 10

Running: testCloseDrawerFromConsolePanelAndOpenFromAnotherPanel
Panel sources was opened.
Drawer panel set to console-view
Console scrollTop: 10
